+++ src/features/inventory/types/index.ts
@@ -1,0 +1,2 @@
+export * from "./dish.types";
+export * from "./ingredient.types";
