+++ src/types/index.ts
@@ -1,6 +1,4 @@
-export * from "../features/inventory/types/dish.types";
-export * from "../features/inventory/types/ingredient.types";
-export * from "../features/menu/types/menu.types";
+
 export * from "../features/cart/types/purchase.types";
 export * from "../features/cart/types/cart.types";
 export * from "../features/cart/types/cart.helpers";
